+++ package.json
@@ -6,9 +6,6 @@
   "private": false,
   "scripts": {
     "postinstall": "husky install",
-    "prepublishOnly": "pinst --disable",
-    "postpublish": "pinst --enable",
-    "prebuild": "rimraf dist",
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
@@ -59,8 +56,6 @@
     "fastify": "^3.28.0",
     "fastify-swagger": "^5.1.1",
     "pg": "^8.7.1",
-    "reflect-metadata": "^0.1.13",
-    "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
     "typeorm": "^0.2.40"
   },
@@ -78,7 +73,6 @@
     "eslint-plugin-prettier": "^4.0.0",
     "husky": "^7.0.4",
     "jest": "^27.2.5",
-    "pinst": "^2.1.6",
     "prettier": "^2.3.2",
     "source-map-support": "^0.5.20",
     "supertest": "^6.1.3",
