--- src/apischema/files.api.shema.ts
@@ -1,11 +1,0 @@
-export const UploadFileSchema = {
-  schema: {
-    type: 'object',
-    properties: {
-      file: {
-        type: 'string',
-        format: 'binary',
-      },
-    },
-  },
-};
