+++ src/controller/index.ts
@@ -1,3 +1,4 @@
 export { SystemController } from './system.controller';
 export { UserController } from './user.controller';
 export { FileManagerController } from './file_manager.controller';
+export { WebsocketController } from './websocket.controller';
