+++ src/controller/websocket.controller.ts
@@ -7,8 +7,9 @@ import {
 } from '@nestjs/websockets';
 import { Logger } from '@nestjs/common';
 import { Socket, Server } from 'socket.io';
-import { UserService } from '../service/user/user.service';
-import { WebsocketService } from '../service/websocket/websocket.service';
+
+import { UserService } from 'src/service/user';
+import { WebsocketService } from 'src/service/websocket';
 
 @WebSocketGateway({
   cors: {
@@ -16,7 +17,7 @@ import { WebsocketService } from '../service/websocket/websocket.service';
   },
   path: '/socket',
 })
-export class WebsocketController implements OnGatewayInit, OnGatewayConnection, OnGatewayDisconnect {
+export class WebsocketController implements OnGatewayInit, OnGatewayConnection {
   constructor(private readonly userService: UserService, private readonly websocketService: WebsocketService) {}
 
   @WebSocketServer() server: Server;
@@ -26,13 +27,7 @@ export class WebsocketController implements OnGatewayInit, OnGatewayConnection, 
     this.websocketService.server = server;
   }
 
-  handleDisconnect(socket: Socket) {
-    this.logger.log(`Client disconnected: ${socket.id}`);
-  }
-
   public async handleConnection(socket: Socket) {
-    this.logger.log(`Client connected: ${socket.id}`);
-
     const accessToken = socket.handshake.auth.accessToken;
     const user = await this.userService.getUserByToken(accessToken);
