+++ src/model/index.ts
@@ -1,2 +1,3 @@
 export { User } from './user';
 export { File } from './file';
+export { Folder } from './folder';
