+++ src/module/index.ts
@@ -1,0 +1,4 @@
+export { FileManagerModule } from './file_manager.module';
+export { SystemModule } from './system.module';
+export { UserModule } from './user.module';
+export { WebsocketModule } from './websocket.module';
