+++ src/service/storage/index.ts
@@ -1,0 +1,1 @@
+export { StorageService } from './storage.service';
