+++ src/service/user/index.ts
@@ -1,0 +1,2 @@
+export { UserFormatter } from './user.formatter';
+export { UserService } from './user.service';
