+++ src/service/user/user.formatter.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 import { UserResponse } from 'src/interface/apiResponse';
-import { User } from 'src/model/user';
+import { User } from 'src/model';
 
 @Injectable()
 export class UserFormatter {
