+++ src/service/websocket/index.ts
@@ -1,0 +1,2 @@
+export { WebsocketEvent } from './events';
+export { WebsocketService } from './websocket.service';
